refactor: combine lib/term
