+++ apps/antalmanac/src/components/RightPane/CoursePane/SearchForm/TermSelector.tsx
@@ -1,6 +1,6 @@
 import { LabeledAutocomplete } from '$components/RightPane/CoursePane/SearchForm/LabeledInputs/LabeledAutocomplete';
 import RightPaneStore, { CourseSearchWarningType } from '$components/RightPane/RightPaneStore';
-import { getTermLongName, termData } from '$lib/termData';
+import { getTermLongName, termData } from '$lib/term';
 import { ComponentProps, useCallback, useEffect, useState } from 'react';
 
 type TermSelectorProps = Omit<
