+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/action-cell/ActionCell.tsx
@@ -6,7 +6,7 @@ import { SectionActionMenu } from '$components/RightPane/SectionTable/SectionTab
 import { TableBodyCellContainer } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/TableBodyCellContainer';
 import { useIsMobile } from '$hooks/useIsMobile';
 import analyticsEnum from '$lib/analytics/analytics';
-import type { Term } from '$lib/termData';
+import type { Term } from '$lib/term';
 import AppStore from '$stores/AppStore';
 import { useNotificationStore } from '$stores/NotificationStore';
 import { Box, CircularProgress, IconButton } from '@mui/material';
