+++ apps/antalmanac/src/stores/NotificationStore.ts
@@ -1,6 +1,6 @@
 import trpc from '$lib/api/trpc';
 import { Notifications } from '$lib/notifications';
-import { buildTermShortName, parseTermShortName, type Quarter } from '$lib/term-constants';
+import { buildTermShortName, parseTermShortName, type Quarter } from '$lib/term';
 import { useSessionStore } from '$stores/SessionStore';
 import { debounce } from '@mui/material';
 import { type AASection, type CourseInfo, WebsocSectionStatusSchema } from '@packages/antalmanac-types';
